+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This file is used to list changes made in each version of grafana.
 
 ## Unreleased
 
+- Added `allow_loading_unsigned_plugins` management
+
 ## 9.5.2 - *2021-01-16*
 
 - Fixed docuemntation showing the incorrect action